refactor(devs-infra): rename util dir to utils dir
