+++ src/ts-jest-transformer.ts
@@ -6,11 +6,11 @@ import { inspect } from 'util'
 import { ConfigSet } from './config/config-set'
 import { JS_JSX_REGEX, TS_TSX_REGEX } from './constants'
 import type { TsJestGlobalOptions } from './types'
-import { stringify } from './util/json'
-import { JsonableValue } from './util/jsonable-value'
-import { rootLogger } from './util/logger'
-import { Errors, interpolate } from './util/messages'
-import { sha1 } from './util/sha1'
+import { stringify } from './utils/json'
+import { JsonableValue } from './utils/jsonable-value'
+import { rootLogger } from './utils/logger'
+import { Errors, interpolate } from './utils/messages'
+import { sha1 } from './utils/sha1'
 
 const INSPECT_CUSTOM = inspect.custom || 'inspect'
