+++ src/utils/get-package-version.ts
