+++ utils/index.js
@@ -1,1 +1,1 @@
-module.exports = require('../dist/util')
+module.exports = require('../dist/utils')
